requirements.txt
undetected-chromedriver==3.5.5
selenium==4.22.0
PySocks==1.7.1
setuptools==70.1.1
beautifulsoup4==4.12.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.0
websockets==12.0
redis==5.0.7
celery==5.4.0
python-socketio==5.11.3
aioredis==2.0.1
aiohttp==3.9.5
googletrans==4.0.0-rc1